# Speed up library page