--- a/app/components/TemplateList.jsx
+++ b/app/components/TemplateList.jsx
@@ -27,7 +27,7 @@ export default class TemplateList extends React.Component {
       let friendlySlug = SlugMaker.make(item.Name);
       return (
         <li className={"item-summary " + (item.ScriptClass || "")} key={index + "." + item.Name}>
-          <img src={item.Logo} loading="lazy" />
+          <img loading="lazy" src={item.Logo} />
           <h4 key={index + "." + item.Name + ".0"}>
             <Link to={`/step-templates/${item.Id}/${friendlySlug}`}>{item.Name}</Link>
           </h4>