We're looking at the loading time issue on library.octopus.com and think we can change the strategy for images to improve things.

This removes the inlining of images and instead refers to images on i.octopus.com.

The only templates affected are the Azure Load Balancer steps, which have specific logos rather than depending on categories.

The screenshot shows these specific images. We could create a category for these if necessary.

**Before** (production)

Server: 3,700 ms
Download: 2,370 ms

**After** (running in a GitHub code space)

Server: 483 ms
Download: 101 ms

## Before

![image](https://github.com/OctopusDeploy/Library/assets/99181436/69c61c10-c438-4729-9716-0eab69ddd653)

![image](https://github.com/OctopusDeploy/Library/assets/99181436/0c72d1e5-4b9a-461e-ba42-df918f1750b5)

## After

![image](https://github.com/OctopusDeploy/Library/assets/99181436/3929c692-d777-4e7d-9f17-fbace196b543)

![image](https://github.com/OctopusDeploy/Library/assets/99181436/7f93d595-0d6a-4bfc-930b-17ac3874a47e)